Do not call `make` in Makefile directly
It should be avoided to call `make` inside a Makefile directly. Otherwise environment variables might not be passed correctly to the submakefile.

For example on BSD you would normally use `gmake` since the build in `make` is not compatible with GNU Make. Hardcoding a call to `make` inside a Makefile will limit the portability.