Enable overwriting of C compiler
When the CC environment variable is set, it will be used as C compiler. Otherwise, `gcc` will be used just as it was the case before.

The reason for this commit is to enable injecting the use of a cross compiler like MXE without symlinking or modification to the Makefile.